Added code to import Libraries.io Java GitHub repository list
In the original paper, it is very strange to re-use an old repo list from an ICPC-2019 paper. In the future, we choose to build dep change seq from a selected repository list based on Libraries.io (I believe it is newer and more comprehensive). This will be the first step to achieve this.

Currently the libraries.io repository database has been already constructed on World of Code MongoDB. Some manual setup will be needed for local construction, but I currently do not have any plan for that.